yc-backend-api/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0
django-cors-headers==4.3.1
psycopg==3.1.13
python-decouple==3.8
dj-database-url==2.1.0
gunicorn==21.2.0

# AI Libraries
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
cohere>=4.0.0

# Development tools
black==23.12.1
flake8==7.0.0